requirements.txt
pandas
numpy
scipy
umap-learn
sklearn
matplotlib
phate
scprep
